# openapi-complete.yaml
openapi: 3.0.3
info:
  title: Sistema de Tutorías API
  description: |
    API REST para el Sistema de Tutorías que facilita la gestión de tutorías entre tutores y tutorados.
    
    El sistema está desarrollado con Spring Boot siguiendo la Arquitectura Hexagonal y permite:
    - Gestión de usuarios (tutores, tutorados, administradores) con integración externa
    - Solicitudes de tutoría con estados avanzados (Conversando, Solicitud_Cancelacion)
    - Asignación y gestión de tutorías con vista detallada
    - Programación de sesiones de tutoría
    - Sistema de retroalimentación con evaluador automático
    - Gestión de habilidades y capítulos
    - Filtros avanzados y búsquedas con operadores OR
    - Autenticación basada en Google ID con contexto de usuario
    - Vista detallada de tutorías con sesiones y feedbacks integrados
  version: 2.0.0
  contact:
    name: Sistema de Tutorías
    email: support@sistematutorias.com
  license:
    name: Propietaria
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo
  - url: http://pragma-tutorias-dev-alb-tutorias-1944959947.us-east-1.elb.amazonaws.com:8080
    description: Servidor de aws
security:
  - GoogleAuth: []
...
paths:
...
  /api/v1/tutorings:
    post:
      tags:
        - Tutorings
      summary: Crear una tutoría
      description: Crea una nueva tutoría asignando un tutor a una solicitud
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTutoringDto'
      responses:
        '201':
          description: Tutoría creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponseTutoring'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: No autorizado - Token de autenticación faltante o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorDto'
        '403':
          description: Prohibido - Usuario no registrado en el sistema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorDto'
    get:
      tags:
        - Tutorings
      summary: Obtener todas las tutorías
      description: |
        Obtiene la lista completa de todas las tutorías en el sistema.
        Retorna información básica de cada tutoría incluyendo tutor, tutorado, 
        habilidades, fechas y estado actual.
      responses:
        '200':
          description: Lista de tutorías obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponseTutoringList'
        '401':
          description: No autorizado - Token de autenticación faltante o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorDto'
        '403':
          description: Prohibido - Usuario no registrado en el sistema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorDto'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /api/v1/tutorings/{tutoringId}/complete:
    patch:
      tags:
        - Tutorings
      summary: Completar una tutoría
      description: Marca una tutoría como completada
      parameters:
        - name: tutoringId
          in: path
          required: true
          schema:
            type: string
          description: ID de la tutoría
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTutoringDto'
      responses:
        '200':
          description: Tutoría completada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponseTutoring'
        '400':
          description: Operación no válida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: No autorizado - Token de autenticación faltante o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorDto'
        '403':
          description: Prohibido - Usuario no registrado en el sistema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorDto'
        '404':
          description: Tutoría no encontrada
  /api/v1/tutorings/{tutoringId}/cancel:
    patch:
      tags:
        - Tutorings
      summary: Cancelar una tutoría
      description: Cancela una tutoría activa
      parameters:
        - name: tutoringId
          in: path
          required: true
          schema:
            type: string
          description: ID de la tutoría
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTutoringStatusDto'
      responses:
        '200':
          description: Tutoría cancelada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponseTutoring'
        '400':
          description: Operación no válida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: No autorizado - Token de autenticación faltante o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorDto'
        '403':
          description: Prohibido - Usuario no registrado en el sistema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorDto'
        '404':
          description: Tutoría no encontrada
...
components:
  securitySchemes:
    GoogleAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Google User ID para autenticación. 
        
        **Formato:** Enviar directamente el Google User ID en el header Authorization.
        
        **Ejemplo:** `Authorization: 1234567890123456789`
        
        **Nota:** En un entorno de producción, esto sería típicamente un JWT token.
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          description: ID único del usuario
        firstName:
          type: string
          description: Nombre del usuario
        lastName:
          type: string
          description: Apellido del usuario
        email:
          type: string
          format: email
          description: Correo electrónico del usuario
        slackId:
          type: string
          description: ID de Slack del usuario
        chapter:
          $ref: '#/components/schemas/ChapterDto'
        rol:
          $ref: '#/components/schemas/RolUsuario'
        activeTutoringLimit:
          type: integer
          description: Límite de tutorías activas
        seniority:
          type: integer
          description: Nivel de seniority del usuario
...
    ChapterDto:
      type: object
      properties:
        id:
          type: string
          description: ID único del capítulo
        name:
          type: string
          description: Nombre del capítulo
...
    SkillDto:
      type: object
      properties:
        id:
          type: string
          description: ID único de la habilidad
        name:
          type: string
          description: Nombre de la habilidad
...
    TutoringDto:
      type: object
      properties:
        id:
          type: string
          description: ID único de la tutoría
        tutor:
          $ref: '#/components/schemas/UserDto'
        tutee:
          $ref: '#/components/schemas/UserDto'
        skills:
          type: array
          items:
            $ref: '#/components/schemas/SkillDto'
        startDate:
          type: string
          format: date-time
          description: Fecha de inicio
        expectedEndDate:
          type: string
          format: date-time
          description: Fecha esperada de finalización
        status:
          $ref: '#/components/schemas/TutoringStatus'
        objectives:
          type: string
          description: Objetivos de la tutoría
    CreateTutoringDto:
      type: object
      required:
        - tutoringRequestId
        - tutorId
        - objectives
      properties:
        tutoringRequestId:
          type: string
          description: ID de la solicitud de tutoría
        tutorId:
          type: string
          description: ID del tutor asignado
        objectives:
          type: string
          description: Objetivos de la tutoría
    UpdateTutoringStatusDto:
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
          description: ID del usuario que realiza la acción
        comments:
          type: string
          description: Comentarios adicionales
    CompleteTutoringDto:
      type: object
      required:
        - userId
        - finalActUrl
      properties:
        userId:
          type: string
          description: ID del usuario que completa la tutoría
        finalActUrl:
          type: string
          description: URL del acta final de la tutoría

    # Tutoring Session schemas
...
    RolUsuario:
      type: string
      enum:
        - TUTOR
        - TUTORADO
        - ADMINISTRADOR
      description: Rol del usuario en el sistema
...
    TutoringStatus:
      type: string
      enum:
        - ACTIVA
        - COMPLETADA
        - CANCELADA
        - SOLICITUD_CANCELACION
      description: |
        Estado de la tutoría:
        - ACTIVA: Tutoría en curso
        - COMPLETADA: Tutoría finalizada exitosamente
        - CANCELADA: Tutoría cancelada
        - SOLICITUD_CANCELACION: Cancelación solicitada pero no aprobada
...
    OkResponseTutoring:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de respuesta
        data:
          $ref: '#/components/schemas/TutoringDto'
        timestamp:
          type: string
          format: date-time
          description: Timestamp de la respuesta
...
    ErrorResponseDto:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de error
        errors:
          type: object
          additionalProperties:
            type: string
          description: Detalles específicos de errores de validación
        timestamp:
          type: string
          format: date-time
          description: Timestamp del error
    AuthenticationErrorDto:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de error de autenticación
          example: "Authorization header is required"
        timestamp:
          type: string
          format: date-time
          description: Timestamp del error
          example: "2024-01-15T10:30:00Z"

    # Statistics schemas
...
    OkResponseTutoringList:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de respuesta
        data:
          type: array
          items:
            $ref: '#/components/schemas/TutoringDto'
        timestamp:
          type: string
          format: date-time
          description: Timestamp de la respuesta
...